+++ magnipy/__init__.py
@@ -1,6 +1,10 @@
-from .distances import *
-from .magnitude import *
-from .function_utils import *
-from .magnitude_dimension import *
-from .summaries import *
+from .magnitude.distances import *
+from .magnitude.compute import *
+from .magnitude.weights import *
+from .magnitude.scales import *
+from .magnitude.convergence import *
+
+# from .magnitude.function_utils import *
+from .magnitude.dimension import *
+from .magnitude.function_operations import *
 from .magnipy import Magnipy
