+++ magnipy/magnipy.py
@@ -1,26 +1,34 @@
-from magnipy.magnitude import (
-    compute_t_conv,
+from magnipy.magnitude.scales import (
     get_scales,
     scale_when_scattered,
     scale_when_almost_scattered,
-    compute_magnitude_until_convergence,
+    cut_ts,
+)
+from magnipy.magnitude.convergence import guess_convergence_scale
+from magnipy.magnitude.weights import (
     magnitude_from_weights,
     similarity_matrix,
 )
-from magnipy.magnitude_dimension import (
+from magnipy.magnitude.compute import (
+    compute_magnitude_until_convergence,
+    compute_t_conv,
+)
+from magnipy.magnitude.dimension import (
     magitude_dimension_profile,
     magnitude_dimension,
     magnitude_dimension_profile_exact,
 )
-from magnipy.distances import get_dist
-from magnipy.summaries import mag_area, mag_diff
-from magnipy.function_utils import (
+from magnipy.magnitude.distances import get_dist
+from magnipy.magnitude.function_operations import (
     diff_of_functions,
     sum_of_functions,
+    cut_until_scale,
+    mag_area,
+    mag_diff,
+)
+from magnipy.utils.plots import (
     plot_magnitude_function,
     plot_magnitude_dimension_profile,
-    cut_until_scale,
-    cut_ts,
 )
 import numpy as np
 import copy
@@ -103,14 +111,20 @@ class Magnipy:
         self._Adj = Adj
         if metric != "precomputed":
             self._X = X
-            self._D = get_dist(
-                X,
-                Adj=self._Adj,
-                p=p,
-                metric=metric,
-                normalise_by_diameter=False,
-                n_neighbors=n_neighbors,
-            )
+
+            def compute_distances(X, X2, Adj=None):
+                return get_dist(
+                    X,
+                    X2=X2,
+                    Adj=Adj,
+                    p=p,
+                    metric=metric,
+                    normalise_by_diameter=False,
+                    n_neighbors=n_neighbors,
+                )
+
+            self._get_dist = compute_distances
+            self._D = self._get_dist(X, X2=None, Adj=self._Adj)
             self._n = self._D.shape[0]
             self._target_value = target_prop * self._D.shape[0]
             self._Z = similarity_matrix(self._D)
@@ -126,16 +140,33 @@ class Magnipy:
             raise Exception(
                 "The scale finding method must be either 'scattered' or 'convergence'."
             )
+
+        def compute_mag(Z, ts, n_ts=n_ts, get_weights=False):
+            return compute_magnitude_until_convergence(
+                Z,
+                ts=ts,
+                n_ts=n_ts,
+                method=method,
+                log_scale=log_scale,
+                get_weights=get_weights,
+                one_point_property=one_point_property,
+                perturb_singularities=perturb_singularities,
+                positive_magnitude=positive_magnitude,
+                input_distances=False,
+            )
+
+        self._compute_mag = compute_mag
+
         self._scale_finding = scale_finding
         self._ts = ts
         self._n_ts = n_ts
         self._log_scale = log_scale
         self._method = method
         self._metric = metric
-        self._p = p
+        # self._p = p
         self._one_point_property = one_point_property
         self._perturb_singularities = perturb_singularities
-        self._n_neighbors = n_neighbors
+        # self._n_neighbors = n_neighbors
         self._return_log_scale = return_log_scale
         self._recompute = recompute
         self._positive_magnitude = positive_magnitude
@@ -151,19 +182,22 @@ class Magnipy:
         self._t_scattered = None
         self._t_almost_scattered = None
 
+    #  ╭──────────────────────────────────────────────────────────╮
+    #  │ Basic Informations                                       │
+    #  ╰──────────────────────────────────────────────────────────╯
+
+    def get_name(self):
+        """
+        Get the name of the Magnipy object.
+        """
+        return self._name
+
     def get_dist(self):
         """
         Compute the distance matrix.
         """
         if (self._D is None) | self._recompute:
-            self._D = get_dist(
-                self._X,
-                Adj=self._Adj,
-                p=self._p,
-                metric=self._metric,
-                normalise_by_diameter=False,
-                n_neighbors=self._n_neighbors,
-            )
+            self._D = self._get_dist(self._X, X2=None, Adj=self._Adj)
         return self._D
 
     def get_similarity_matrix(self):
@@ -174,133 +208,9 @@ class Magnipy:
             self._Z = similarity_matrix(self._D)
         return self._Z
 
-    def get_name(self):
-        """
-        Get the name of the Magnipy object.
-        """
-        return self._name
-
-    def get_magnitude_weights(self):
-        """
-        Compute the magnitude weights.
-
-        Returns
-        -------
-        weights : array_like, shape (`n_obs`, `n_ts`)
-            The weights of the magnitude function.
-        """
-        if (self._weights is None) | self._recompute:
-            ts = self.get_scales()
-            weights, ts = compute_magnitude_until_convergence(
-                self._Z,
-                ts=self._ts,
-                n_ts=self._n_ts,
-                method=self._method,
-                log_scale=self._log_scale,
-                get_weights=True,
-                one_point_property=self._one_point_property,
-                perturb_singularities=self._perturb_singularities,
-                positive_magnitude=self._positive_magnitude,
-                input_distances=False,
-            )
-            self._weights = weights
-            self._ts = ts
-            if self._ts is None:
-                self._t_conv = ts[-1]
-        return self._weights, self._ts
-
-    def get_magnitude(self):
-        """
-        Compute the magnitude function.
-
-        Returns
-        -------
-        magnitude : array_like, shape (`n_ts`, )
-            The values of the magnitude function.
-        ts : array_like, shape (`n_ts`, )
-            The scales at which the magnitude function has been evaluated.
-        """
-        if ((self._magnitude is None) & (self._weights is None)) | self._recompute:
-            ts = self.get_scales()
-            self._magnitude, ts = compute_magnitude_until_convergence(
-                self._Z,
-                ts=self._ts,
-                n_ts=self._n_ts,
-                method=self._method,
-                log_scale=self._log_scale,
-                get_weights=False,
-                one_point_property=self._one_point_property,
-                perturb_singularities=self._perturb_singularities,
-                positive_magnitude=self._positive_magnitude,
-                input_distances=False,
-            )
-            if self._ts is None:
-                self._t_conv = ts[-1]
-                self._ts = ts
-        elif (self._magnitude is None) & (not (self._weights is None)):
-            self._magnitude = magnitude_from_weights(self._weights)
-        return self._magnitude, self._ts
-
-    def plot_magnitude_function(self):
-        """
-        Plot the magnitude function.
-        """
-        if (self._magnitude is None) | self._recompute:
-            _, _ = self.get_magnitude()
-        plot_magnitude_function(self._ts, self._magnitude, name=self._name)
-
-    def get_magnitude_dimension_profile(self, exact=False, h=None):
-        """
-        Compute the magnitude dimension profile.
-
-        Parameters
-        ----------
-        exact : bool
-            Whether to compute the magnitude dimension profile exactly.
-        h : float
-            The stepsize to use for exact computations of the slope.
-        """
-        if (self._magnitude_dimension_profile is None) | self._recompute:
-            if exact:
-                self._magnitude_dimension_profile, self._ts_dim = (
-                    magnitude_dimension_profile_exact(
-                        self._D,
-                        ts=self._ts,
-                        h=h,
-                        target_value=self._target_value,
-                        n_ts=self._n_ts,
-                        return_log_scale=self._return_log_scale,
-                        one_point_property=self._one_point_property,
-                        method=self._method,
-                        log_scale=self._log_scale,
-                    )
-                )
-            else:
-                if self._magnitude is None:
-                    _, _ = self.get_magnitude()
-                (
-                    self._magnitude_dimension_profile,
-                    self._ts_dim,
-                ) = magitude_dimension_profile(
-                    mag=self._magnitude,
-                    ts=self._ts,
-                    return_log_scale=self._return_log_scale,
-                    one_point_property=self._one_point_property,
-                )
-        return self._magnitude_dimension_profile, self._ts_dim
-
-    def plot_magnitude_dimension_profile(self):
-        """
-        Plot the magnitude dimension profile.
-        """
-        if (self._magnitude_dimension_profile is None) | self._recompute:
-            _, _ = self.get_magnitude_dimension_profile()
-        plot_magnitude_dimension_profile(
-            ts=self._ts_dim,
-            mag_dim=self._magnitude_dimension_profile,
-            log_scale=self._return_log_scale,
-            name=self._name,
-        )
+    #  ╭──────────────────────────────────────────────────────────╮
+    #  │ Find the Evaluation Scales                               │
+    #  ╰──────────────────────────────────────────────────────────╯
 
     def get_t_conv(self):
         """
@@ -308,13 +218,24 @@ class Magnipy:
         """
         if self._scale_finding == "convergence":
             if (self._t_conv is None) | self._recompute:
-                self._t_conv = compute_t_conv(
-                    self._Z,
+
+                def comp_mag(X, ts):
+                    return self._compute_mag(X, ts)[0]
+
+                self._t_conv = guess_convergence_scale(
+                    D=self._Z,
+                    comp_mag=comp_mag,
                     target_value=self._target_value,
-                    method=self._method,
-                    positive_magnitude=self._positive_magnitude,
-                    input_distances=False,
+                    guess=10,
                 )
+                # print(f"Convergence scale: {self.t_conv}")
+                # self._t_conv = compute_t_conv(
+                #    self._Z,
+                #   target_value=self._target_value,
+                #    method=self._method,
+                #    positive_magnitude=self._positive_magnitude,
+                #    input_distances=False,
+                # )
             return self._t_conv
         elif self._scale_finding == "scattered":
             return self._scale_when_almost_scattered(q=None)
@@ -375,6 +296,101 @@ class Magnipy:
         self._weights = None
         self._ts_dim = None
 
+    def _scale_when_scattered(self):
+        """
+        Compute the scale at which the metric space is scattered.
+
+        Returns
+        -------
+        t_scattered : float
+            The scale at which the metric space is scattered.
+
+        References
+        ----------
+        [1] Leinster T. The magnitude of metric spaces. Documenta Mathematica. 2013 Jan 1;18:857-905.
+        """
+        if (self._t_scattered is None) | self._recompute:
+            self._t_scattered = scale_when_scattered(self._D)
+        return self._t_scattered
+
+    def _scale_when_almost_scattered(self, q=None):
+        """
+        Compute the scale at which the metric space is almost scattered.
+
+        Parameters
+        ----------
+        q : float
+            The proportion of points that are scattered.
+
+        Returns
+        -------
+        t_almost_scattered : float
+            The scale at which the metric space is almost scattered.
+        """
+
+        if (self._t_almost_scattered is None) | self._recompute:
+            self._t_almost_scattered = scale_when_almost_scattered(
+                self._D, n=self._n, q=q
+            )
+        return self._t_almost_scattered
+
+    #  ╭──────────────────────────────────────────────────────────╮
+    #  │ Compute Magnitude Weights and Functions                  │
+    #  ╰──────────────────────────────────────────────────────────╯
+
+    def get_magnitude_weights(self):
+        """
+        Compute the magnitude weights.
+
+        Returns
+        -------
+        weights : array_like, shape (`n_obs`, `n_ts`)
+            The weights of the magnitude function.
+        """
+        if (self._weights is None) | self._recompute:
+            ts = self.get_scales()
+            weights, ts = self._compute_mag(
+                self._Z,
+                ts=ts,
+                # n_ts=self._n_ts,
+                # method=self._method,
+                # log_scale=self._log_scale,
+                get_weights=True,
+            )
+            self._weights = weights
+            self._ts = ts
+            if self._ts is None:
+                self._t_conv = ts[-1]
+        return self._weights, self._ts
+
+    def get_magnitude(self):
+        """
+        Compute the magnitude function.
+
+        Returns
+        -------
+        magnitude : array_like, shape (`n_ts`, )
+            The values of the magnitude function.
+        ts : array_like, shape (`n_ts`, )
+            The scales at which the magnitude function has been evaluated.
+        """
+        if ((self._magnitude is None) & (self._weights is None)) | self._recompute:
+            ts = self.get_scales()
+            self._magnitude, ts = self._compute_mag(
+                self._Z,
+                ts=ts,
+                # n_ts=self._n_ts,
+                # method=self._method,
+                # log_scale=self._log_scale,
+                get_weights=False,
+            )
+            if self._ts is None:
+                self._t_conv = ts[-1]
+                self._ts = ts
+        elif (self._magnitude is None) & (not (self._weights is None)):
+            self._magnitude = magnitude_from_weights(self._weights)
+        return self._magnitude, self._ts
+
     def _eval_at_scales(self, ts_new, get_weights=False):
         """
         Evaluate the magnitude functions at new scales.
@@ -393,45 +409,57 @@ class Magnipy:
         ts : array_like, shape (`n_ts_new`, )
             The new scales at which the magnitude function has been evaluated.
         """
-        mag, ts = compute_magnitude_until_convergence(
+        mag, ts = self._compute_mag(
             self._Z,
             ts=ts_new,
-            method=self._method,
             get_weights=get_weights,
-            one_point_property=self._one_point_property,
-            perturb_singularities=self._perturb_singularities,
-            positive_magnitude=self._positive_magnitude,
-            input_distances=False,
         )
         return mag, ts
 
-    def _cut_until_scale(self, t_cut):
+    #  ╭──────────────────────────────────────────────────────────╮
+    #  │ Magnitude Dimension                                      │
+    #  ╰──────────────────────────────────────────────────────────╯
+
+    def get_magnitude_dimension_profile(self, exact=False, h=None):
         """
-        Cut the magnitude functions at a given scale.
+        Compute the magnitude dimension profile.
 
         Parameters
         ----------
-        t_cut : float
-            The scale at which to cut the magnitude functions.
+        exact : bool
+            Whether to compute the magnitude dimension profile exactly.
+        h : float
+            The stepsize to use for exact computations of the slope.
         """
-        if self._magnitude is not None:
-            self._magnitude, self._ts = cut_until_scale(
-                self._ts, self._magnitude, t_cut=t_cut, D=self._D, method=self._method
-            )
-        elif self._ts is not None:
-            self._ts = cut_ts(self._ts, t_cut)
-        self._magnitude_area = None
-        self._magnitude_dimension = None
-        if self._magnitude_dimension_profile is not None:
-            self._magnitude_dimension_profile, self._ts_dim = cut_until_scale(
-                self._ts_dim,
-                self._magnitude_dimension_profile,
-                t_cut=t_cut,
-                D=None,
-                method=self._method,
-            )
-        if self._weights is not None:
-            self._weights = self._weights[: len()]
+        if (self._magnitude_dimension_profile is None) | self._recompute:
+            if exact:
+                (
+                    self._magnitude_dimension_profile,
+                    self._ts_dim,
+                ) = magnitude_dimension_profile_exact(
+                    self._D,
+                    ts=self._ts,
+                    h=h,
+                    target_value=self._target_value,
+                    n_ts=self._n_ts,
+                    return_log_scale=self._return_log_scale,
+                    one_point_property=self._one_point_property,
+                    method=self._method,
+                    log_scale=self._log_scale,
+                )
+            else:
+                if self._magnitude is None:
+                    _, _ = self.get_magnitude()
+                (
+                    self._magnitude_dimension_profile,
+                    self._ts_dim,
+                ) = magitude_dimension_profile(
+                    mag=self._magnitude,
+                    ts=self._ts,
+                    return_log_scale=self._return_log_scale,
+                    one_point_property=self._one_point_property,
+                )
+        return self._magnitude_dimension_profile, self._ts_dim
 
     def get_magnitude_dimension(self, exact=False):
         """
@@ -445,7 +473,8 @@ class Magnipy:
         Returns
         -------
         magnitude_dimension : float
-            The magnitude dimension. We compute it as the maximum value of the magnitude dimension profile.
+            The magnitude dimension. We compute it as the maximum value of the
+            magnitude dimension profile.
         """
         if self._magnitude_dimension_profile is None:
             _, _ = self.get_magnitude_dimension_profile(exact=exact)
@@ -454,6 +483,35 @@ class Magnipy:
                 self._magnitude_dimension_profile
             )
         return self._magnitude_dimension
+
+    #  ╭──────────────────────────────────────────────────────────╮
+    #  │ Plots                                                    │
+    #  ╰──────────────────────────────────────────────────────────╯
+
+    def plot_magnitude_function(self):
+        """
+        Plot the magnitude function.
+        """
+        if (self._magnitude is None) | self._recompute:
+            _, _ = self.get_magnitude()
+        plot_magnitude_function(self._ts, self._magnitude, name=self._name)
+
+    def plot_magnitude_dimension_profile(self):
+        """
+        Plot the magnitude dimension profile.
+        """
+        if (self._magnitude_dimension_profile is None) | self._recompute:
+            _, _ = self.get_magnitude_dimension_profile()
+        plot_magnitude_dimension_profile(
+            ts=self._ts_dim,
+            mag_dim=self._magnitude_dimension_profile,
+            log_scale=self._return_log_scale,
+            name=self._name,
+        )
+
+    #  ╭──────────────────────────────────────────────────────────╮
+    #  │ Operations with Magnipy Objects                          │
+    #  ╰──────────────────────────────────────────────────────────╯
 
     def include_points(self, X_new, Adj_new=None, update_ts=False):
         if self._X is None:
@@ -538,7 +596,8 @@ class Magnipy:
                 D = np.delete(D, ind_delete, axis=1)
             self._D = D
             self._Z = similarity_matrix(self._D)
-            # self._D = get_dist(X, p=self._p, metric=self._metric, normalise_by_diameter=False, n_neighbors=self._n_neighbors)
+            # self._D = get_dist(X, p=self._p, metric=self._metric, normalise_by_diameter=False,
+            # n_neighbors=self._n_neighbors)
             self._n = self._D.shape[0]
         if update_ts:
             self._ts = None
@@ -553,15 +612,43 @@ class Magnipy:
         self._weights = None
         self._ts_dim = None
 
+    def _cut_until_scale(self, t_cut):
+        """
+        Cut the magnitude functions at a given scale.
+
+        Parameters
+        ----------
+        t_cut : float
+            The scale at which to cut the magnitude functions.
+        """
+        if self._magnitude is not None:
+            self._magnitude, self._ts = cut_until_scale(
+                self._ts, self._magnitude, t_cut=t_cut, D=self._D, method=self._method
+            )
+        elif self._ts is not None:
+            self._ts = cut_ts(self._ts, t_cut)
+        self._magnitude_area = None
+        self._magnitude_dimension = None
+        if self._magnitude_dimension_profile is not None:
+            self._magnitude_dimension_profile, self._ts_dim = cut_until_scale(
+                self._ts_dim,
+                self._magnitude_dimension_profile,
+                t_cut=t_cut,
+                D=None,
+                method=self._method,
+            )
+        if self._weights is not None:
+            self._weights = self._weights[: len()]
+
     def copy(self):
         """
         Return a copy of the Magnipy object.
         """
         return copy.deepcopy(self)
 
-    def _substract(self, other, t_cut=None, exact=True):
+    def _subtract(self, other, t_cut=None, exact=True):
         """
-        Substract the magnitude functions of two Magnipy objects.
+        Subtract the magnitude functions of two Magnipy objects.
 
         Parameters
         ----------
@@ -579,7 +666,7 @@ class Magnipy:
         """
         if self._metric != other._metric:
             raise Exception(
-                "Magnitude functions need to share the same notion of distance in order to be substracted across the same scales of t!!"
+                "Magnitude functions need to share the same notion of distance in order to be subtracted across the same scales of t!!"
             )
         combined = Magnipy(None)
         combined._magnitude, combined._ts = diff_of_functions(
@@ -632,6 +719,10 @@ class Magnipy:
         )
         combined._n_ts = len(combined._ts)
         return combined
+
+    #  ╭──────────────────────────────────────────────────────────╮
+    #  │ Diversity Summaries                                      │
+    #  ╰──────────────────────────────────────────────────────────╯
 
     def MagArea(
         self,
@@ -736,41 +827,3 @@ class Magnipy:
             name=self._name + " - " + other._name,
         )
         return mag_difference
-
-    def _scale_when_scattered(self):
-        """
-        Compute the scale at which the metric space is scattered.
-
-        Returns
-        -------
-        t_scattered : float
-            The scale at which the metric space is scattered.
-
-        References
-        ----------
-        [1] Leinster T. The magnitude of metric spaces. Documenta Mathematica. 2013 Jan 1;18:857-905.
-        """
-        if (self._t_scattered is None) | self._recompute:
-            self._t_scattered = scale_when_scattered(self._D)
-        return self._t_scattered
-
-    def _scale_when_almost_scattered(self, q=None):
-        """
-        Compute the scale at which the metric space is almost scattered.
-
-        Parameters
-        ----------
-        q : float
-            The proportion of points that are scattered.
-
-        Returns
-        -------
-        t_almost_scattered : float
-            The scale at which the metric space is almost scattered.
-        """
-
-        if (self._t_almost_scattered is None) | self._recompute:
-            self._t_almost_scattered = scale_when_almost_scattered(
-                self._D, n=self._n, q=q
-            )
-        return self._t_almost_scattered
