+++ magnipy/magnitude/dimension.py
@@ -1,11 +1,11 @@
 import numpy as np
-from magnipy.magnitude import (
+from magnipy.magnitude.compute import (
     compute_t_conv,
-    get_scales,
     magnitude_from_distances,
     compute_magnitude_until_convergence,
 )
-from magnipy.distances import get_dist
+from magnipy.magnitude.scales import get_scales
+from magnipy.magnitude.distances import get_dist
 
 
 def compute_magnitude_dimension_profile(
@@ -21,19 +21,23 @@ def compute_magnitude_dimension_profile(
     normalise_by_diameter=False,
     one_point_property=False,
     n_neighbors=12,
-    exact=False,
+    exact=False,  ### TODO: rename
     return_log_scale=False,
+    input_distances=True,
 ):
     """
     Compute the magnitude dimension profile of a dataset X.
     """
-    D = get_dist(
-        X,
-        p=p,
-        metric=metric,
-        normalise_by_diameter=normalise_by_diameter,
-        n_neighbors=n_neighbors,
-    )
+    if input_distances:
+        D = X
+    else:
+        D = get_dist(
+            X,
+            p=p,
+            metric=metric,
+            normalise_by_diameter=normalise_by_diameter,
+            n_neighbors=n_neighbors,
+        )
     if exact:
         slopes, ts = magnitude_dimension_profile_exact(
             D,
